Modularized user related & database functions
